LLM/llm_cuda121_final_requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
accelerate==0.34.0
datasets==3.0.0
deepspeed==0.14.0
sentencepiece==0.2.0
torch==2.2.1+cu121
transformers==4.45.0
tqdm==4.66.5
PyYAML==6.0.1
tensorboard==2.15.1
numpy==1.26.4
pandas==2.2.2
trl==0.11.4
peft==0.13.21